requirements.txt
astroid==2.4.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pbkdf2==1.3
pyaes==1.6.1
pycryptodome==3.9.8
pylint==2.5.3
six==1.15.0
toml==0.10.1
wrapt==1.12.1
